api/requirements.txt
fastapi>=0.108.0
python-multipart>=0.0.6

pydantic>=2.6
pydantic_core>=2.14
pydantic-settings>=2.1

SQLAlchemy[asyncio]>=2.0
sqlalchemy-utils>=0.41
types-sqlalchemy-utils
sqlalchemy-file>=0.6
fasteners>=0.19
psycopg[c]>=3.1
alembic>=1.12
uvicorn[standard]>=0.23

passlib>=1.7
python-jose>=3.3
types-python-jose
types-passlib
pyhumps>=3.8

aiofiles>=23.2.1
types-aiofiles>=23.2.0

pandas>=2.1.4
openpyxl>=3.1.2
pandas-stubs

watchfiles==0.24.0

# DEV ONLY

packaging>=23.2
mypy>=1.6.0
debugpy>=1.8.0
typing_extensions>=4.9.0
black>=23.11
